Refactor the test app without solid-ui-react
Move some components around, remove all of the styling, and simplify the code going directly to solid-client-authn-browser.
