+++ e2e/browser/testApp/package.json
@@ -19,7 +19,6 @@
     "@inrupt/solid-client": "^1.20.1",
     "@inrupt/solid-client-access-grants": "file:../../../",
     "@inrupt/solid-client-authn-browser": "^1.12.2",
-    "@inrupt/solid-ui-react": "^2.7.0",
     "next": "^12.1.0",
     "react": "^16.14.0",
     "react-dom": "^16.14.0"
